moved Modal.tsx to shared: Ignored ESLint sistent warnings
